Restructure project layout
Restructured project layout to allow using the lib via
CMakes `add_subdirectory`.
